requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pandas==1.3.3
reportlab==3.6.3
xlsxwriter==3.0.2
Werkzeug==2.0.1
python-dotenv==0.19.0